requirements.txt
absl-py==0.9.0
antlr4-python3-runtime==4.7.2
astor==0.8.1
astroid==2.1.0
beautifulsoup4==4.7.1
cachetools==4.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.9
dominate==2.4.0
en-core-web-sm==2.0.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
mysql-connector-python==8.0.17
nltk==3.4.5
numpy==1.16.1
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
protobuf==3.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.2.2
radb==3.0.2
regex==2018.1.10
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.4.1
six==1.12.0
soupsieve==1.9
spacy==2.0.18
SQLAlchemy==1.3.3
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.31.1
ujson==1.35
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
